Integrate introduced Object Store code to BulkPrepareHandler code
