+++ src/handlers/lib/utility.js
@@ -236,7 +236,7 @@ const getKafkaConfig = (flow, functionality, action) => {
     actionObject.config.logger = Logger
     return actionObject.config
   } catch (e) {
-    throw new Error(`No config found for those parameters flow='${flow}', functionality='${functionality}', action='${action}'`)
+    throw new Error(`No config found for flow='${flow}', functionality='${functionality}', action='${action}'`)
   }
 }
